Further reorganized userspace to make more modular
