+++ users/drashna/drashna.h
@@ -20,6 +20,8 @@
 #include "eeprom.h"
 #include "keyrecords/wrappers.h"
 #include "keyrecords/process_records.h"
+#include "callbacks.h"
+
 #ifdef TAP_DANCE_ENABLE
 #    include "keyrecords/tap_dances.h"
 #endif  // TAP_DANCE_ENABLE
@@ -58,7 +60,7 @@ enum userspace_layers {
     _ADJUST,
 };
 
-#define _MACROS          _MOUSE
+#define _MACROS _MOUSE
 #define _DEFAULT_LAYER_1 FIRST_DEFAULT_LAYER
 #define _DEFAULT_LAYER_2 (FIRST_DEFAULT_LAYER + 1)
 #define _DEFAULT_LAYER_3 (FIRST_DEFAULT_LAYER + 2)
@@ -86,23 +88,10 @@ enum userspace_layers {
 #define DEFAULT_LAYER_3_RGB RGB_MAGENTA
 #define DEFAULT_LAYER_4_RGB RGB_GOLDENROD
 
-bool          mod_key_press_timer(uint16_t code, uint16_t mod_code, bool pressed);
-bool          mod_key_press(uint16_t code, uint16_t mod_code, bool pressed, uint16_t this_timer);
-void          matrix_init_keymap(void);
-void          matrix_init_secret(void);
-void          shutdown_keymap(void);
-void          suspend_power_down_keymap(void);
-void          suspend_wakeup_init_keymap(void);
-void          matrix_scan_keymap(void);
-void          matrix_scan_secret(void);
-layer_state_t layer_state_set_keymap(layer_state_t state);
-layer_state_t default_layer_state_set_keymap(layer_state_t state);
-void          led_set_keymap(uint8_t usb_led);
-void          eeconfig_init_keymap(void);
-bool          hasAllBitsInMask(uint8_t value, uint8_t mask);
-#ifdef SPLIT_KEYBOARD
-void matrix_slave_scan_keymap(void);
-#endif
+bool mod_key_press_timer(uint16_t code, uint16_t mod_code, bool pressed);
+bool mod_key_press(uint16_t code, uint16_t mod_code, bool pressed, uint16_t this_timer);
+bool hasAllBitsInMask(uint8_t value, uint8_t mask);
+void tap_code16_nomods(uint8_t kc);
 
 // clang-format off
 typedef union {
@@ -119,20 +108,3 @@ typedef union {
 // clang-format on
 
 extern userspace_config_t userspace_config;
-
-/*
-Custom Keycodes for Diablo 3 layer
-But since TD() doesn't work when tap dance is disabled
-We use custom codes here, so we can substitute the right stuff
-*/
-#ifdef TAP_DANCE_ENABLE
-#    define KC_D3_1 TD(TD_D3_1)
-#    define KC_D3_2 TD(TD_D3_2)
-#    define KC_D3_3 TD(TD_D3_3)
-#    define KC_D3_4 TD(TD_D3_4)
-#else  // TAP_DANCE_ENABLE
-#    define KC_D3_1 KC_1
-#    define KC_D3_2 KC_2
-#    define KC_D3_3 KC_3
-#    define KC_D3_4 KC_4
-#endif  // TAP_DANCE_ENABLE
